refactor(ui): move box geometry and wrapping math into ui/kit
The titled and bilateral border builders, the three wrap flavors, the
ANSI stripper, the blank-row trimmers and the visible-width measurement
move verbatim into ui/kit; the agent renderer and the response envelope
keep their exact signatures as thin delegates, so every historical call
site and test keeps working while chat, agent and command surfaces now
share one geometry.

The five divergent terminal-width helpers now delegate to the kit's
single one. The only behavior change is deliberate: off-terminal runs
(pipes, CI) fall back to 100 columns everywhere instead of a mix of 80
and 100, and truncation only appends a color reset on profiles that
emit color at all. Width-sensitive tests updated to assert against the
unified helper; the runewidth emoji-width pin moved with the geometry
and keeps its guard test.
