+++ cli/agent/ui_renderer.go
@@ -18,9 +18,9 @@ import (
 
 	"github.com/charmbracelet/lipgloss"
 	"github.com/diillson/chatcli/i18n"
+	"github.com/diillson/chatcli/ui/kit"
 	"github.com/diillson/chatcli/ui/theme"
 	"go.uber.org/zap"
-	"golang.org/x/term"
 )
 
 // typewriterPrint emits text rune-by-rune with a small delay between
@@ -152,10 +152,7 @@ func (r *UIRenderer) StreamOutput(line string) {
 	// (ex.: `kubectl ... -o yaml` com annotations gigantes / blocos
 	// `last-applied-configuration`). Sem isso, uma linha que estoura a
 	// largura do terminal faz reflow e rasga a borda lateral/rodapé do box.
-	termWidth, err := terminalWidthForStream()
-	if err != nil || termWidth <= 0 {
-		termWidth = 80
-	}
+	termWidth := kit.TermWidth()
 
 	// Cada linha emitida é: prefix("│  ") + icon + conteúdo. A largura
 	// visível total precisa caber em termWidth-2 (mesma gutter de 2 cols
@@ -181,63 +178,10 @@ func (r *UIRenderer) StreamOutput(line string) {
 	}
 }
 
-// terminalWidthForStream retorna a largura atual do terminal (cols).
-func terminalWidthForStream() (int, error) {
-	w, _, err := term.GetSize(int(os.Stdout.Fd())) //#nosec G115 -- value bounded by domain
-	return w, err
-}
-
 // wrapStreamLine quebra UMA linha de output cru de tool na largura visível
-// informada. Diferente de wrapText, NÃO colapsa espaços em branco: a
-// indentação inicial é preservada (e repetida nas continuações) para que
-// YAML/JSON estruturado continue legível dentro do box. A quebra é por
-// runa (ANSI/wide-rune aware via lipgloss.Width), evitando cortar
-// sequências multibyte no meio.
+// informada, preservando indentação. Matemática em kit.WrapStreamLine.
 func wrapStreamLine(line string, width int) []string {
-	if width <= 0 || VisibleLen(line) <= width {
-		return []string{line}
-	}
-
-	trimmed := strings.TrimLeft(line, " \t")
-	indent := line[:len(line)-len(trimmed)]
-	indentW := VisibleLen(indent)
-	if indentW >= width-1 {
-		// Indentação maior que o box: desiste de preservá-la.
-		indent = ""
-		indentW = 0
-	}
-	chunkW := width - indentW
-	if chunkW < 1 {
-		chunkW = 1
-	}
-
-	var out []string
-	var cur strings.Builder
-	curW := 0
-	flush := func() {
-		out = append(out, indent+cur.String())
-		cur.Reset()
-		curW = 0
-	}
-
-	for _, rr := range trimmed {
-		rw := lipgloss.Width(string(rr))
-		if rw < 1 {
-			rw = 1
-		}
-		if curW+rw > chunkW && cur.Len() > 0 {
-			flush()
-		}
-		cur.WriteRune(rr)
-		curW += rw
-	}
-	if cur.Len() > 0 {
-		flush()
-	}
-	if len(out) == 0 {
-		out = append(out, indent)
-	}
-	return out
+	return kit.WrapStreamLine(line, width)
 }
 
 // Colorize aplica cores ANSI (exportada com C maiúsculo). O resultado passa
@@ -521,18 +465,12 @@ func (r *UIRenderer) PrintPrompt() string {
 	return r.Colorize(i18n.T("agent.prompt.choice"), ColorLime)
 }
 
-// VisibleLen calcula comprimento visível em colunas do terminal (sem ANSI codes).
-//
-// Delegates to lipgloss.Width — the same function the bordered box
-// renderer uses to size top/bottom borders. Sharing one measurement
-// path is what keeps wrap math and border math in agreement when the
-// content has emoji presentation sequences (e.g. "🏟️" = stadium + VS-16
-// = U+1F3DF + U+FE0F) that pure runewidth.StringWidth reports as 1 col
-// while every modern terminal renders as 2. Mismatch there used to
-// drift the right border outside the visible box; routing both sides
-// through lipgloss.Width removes the disagreement entirely.
+// VisibleLen calcula comprimento visível em colunas do terminal (sem ANSI
+// codes). Delegates to kit.VisibleLen (lipgloss.Width) — one measurement
+// path for wrap math and border math keeps them in agreement when content
+// has emoji presentation sequences.
 func VisibleLen(s string) int {
-	return lipgloss.Width(s)
+	return kit.VisibleLen(s)
 }
 
 // RenderTimelineEvent desenha um "card" estilizado com:
@@ -568,14 +506,9 @@ func (r *UIRenderer) RenderAssistantResponseTimelineEvent(icon, title, content, 
 // tool output (YAML/JSON/tables, where indentation and column alignment must
 // survive).
 func (r *UIRenderer) renderTimelineEventInner(icon, title, content, color string, typewrite bool, wrapFn func(string, int) []string) {
-	termWidth, _, err := term.GetSize(int(os.Stdout.Fd())) //#nosec G115 -- value bounded by domain
-	if err != nil || termWidth <= 0 {
-		termWidth = 80
-	}
-
-	// Reserve 2 cols on the right edge for terminals with native
+	// kit.ContentWidth already reserves the right-edge gutter for native
 	// scrollbars (iTerm, VSCode) so the box border never gets clipped.
-	maxCardWidth := termWidth - 2
+	maxCardWidth := kit.ContentWidth()
 	if maxCardWidth < 24 {
 		maxCardWidth = 24
 	}
@@ -645,90 +578,21 @@ func (r *UIRenderer) renderTimelineEventInner(icon, title, content, color string
 //   - title longer than card: truncate dashes to fit (header may overflow
 //     by 1-2 cols on extreme widths; acceptable degradation)
 func buildTitledTopBorder(header string, targetWidth int, color string, r *UIRenderer) string {
-	// Visible cols reserved: `╭── ` (4) + header (lipgloss) + ` ` (1) + dashes + `╮` (1)
-	usedNoFill := 4 + lipgloss.Width(header) + 1 + 1
-	fill := targetWidth - usedNoFill
-	if fill < 0 {
-		// Header doesn't fit — emit a minimal top without filling.
-		// The card will still close at the right cardWidth because we
-		// honor lipgloss's body measurement.
-		return r.Colorize("╭── "+header+" ╮", color+ColorBold)
-	}
-	line := "╭── " + header + " " + strings.Repeat("─", fill) + "╮"
-	return r.Colorize(line, color+ColorBold)
+	// Geometry in kit.TitledTopBorder; coloring stays here so the legacy
+	// ANSI-constant call sites keep byte-identical output.
+	return r.Colorize(kit.TitledTopBorder(header, targetWidth), color+ColorBold)
 }
 
-// trimBlankBoxBodyRows removes fully-empty content rows directly
-// adjacent to the top or bottom border of a lipgloss-rendered box.
-// An empty row looks like "│         │" — same width as the sides
-// but zero printable content between them. The Padding(0,2) on the
-// body style already adds breathing space, so additional blank rows
-// from glamour or wrapped paragraphs would stack up and break the
-// "card opens on real content" invariant.
+// trimBlankBoxBodyRows removes empty body rows adjacent to the box borders.
+// Logic in kit.TrimBlankBoxBodyRows.
 func trimBlankBoxBodyRows(rendered string) string {
-	rows := strings.Split(rendered, "\n")
-	if len(rows) <= 2 {
-		return rendered
-	}
-	isBlankRow := func(s string) bool {
-		plain := stripANSIForCard(s)
-		// Empty body rows are "│  ...spaces...  │". Anything else
-		// counts as real content (including a ╰── bottom border).
-		if !strings.HasPrefix(plain, "│") || !strings.HasSuffix(plain, "│") {
-			return false
-		}
-		inner := strings.TrimSuffix(strings.TrimPrefix(plain, "│"), "│")
-		return strings.TrimSpace(inner) == ""
-	}
-	// We do NOT touch the first row (it could be a border) or the
-	// last (always the bottom border). Trim only the body slice in
-	// between.
-	start := 0
-	end := len(rows)
-	// Find content boundaries: skip leading empties, then trailing.
-	for start < end && rows[start] != "" && !isBlankRow(rows[start]) {
-		// Non-blank, real row — leave start where it is.
-		break
-	}
-	// Trim trailing empty body rows that sit right before the bottom border.
-	// Bottom border is always the last line (no trailing newline preserved).
-	bottomIdx := end - 1
-	cut := bottomIdx - 1
-	for cut > start && isBlankRow(rows[cut]) {
-		cut--
-	}
-	if cut+1 < bottomIdx {
-		rows = append(rows[:cut+1], rows[bottomIdx])
-	}
-	// Trim leading empty body rows that sit right after the (absent) top
-	// border. Since BorderTop is false, rows[0] is the first content row.
-	leading := 0
-	for leading < len(rows)-1 && isBlankRow(rows[leading]) {
-		leading++
-	}
-	if leading > 0 {
-		rows = rows[leading:]
-	}
-	return strings.Join(rows, "\n")
+	return kit.TrimBlankBoxBodyRows(rendered)
 }
 
-// stripANSIForCard removes CSI color escapes so width and emptiness
-// checks see plain text. Inlined to avoid a regex dependency on the
-// hot card-rendering path.
+// stripANSIForCard removes CSI color escapes so width and emptiness checks
+// see plain text. Logic in kit.StripANSI.
 func stripANSIForCard(s string) string {
-	var b strings.Builder
-	b.Grow(len(s))
-	for i := 0; i < len(s); i++ {
-		if s[i] == 0x1b && i+1 < len(s) && s[i+1] == '[' {
-			i += 2
-			for i < len(s) && (s[i] < 0x40 || s[i] > 0x7e) {
-				i++
-			}
-			continue
-		}
-		b.WriteByte(s[i])
-	}
-	return b.String()
+	return kit.StripANSI(s)
 }
 
 // ansiColorToLip maps the package-local ANSI color constants used by
@@ -751,273 +615,41 @@ func (r *UIRenderer) RenderMarkdownTimelineEvent(icon, title, renderedMarkdownAN
 	r.renderTimelineEventInner(icon, title, renderedMarkdownANSI, color, false, wrapStructured)
 }
 
-// trimBlankBorderRows drops fully-blank rows from the leading and
-// trailing edges of a wrapped-text slice. A row is "blank" when it
-// has zero visible width — color codes alone don't count as visible
-// content (lipgloss/glamour emit ANSI-only lines for some markdown
-// constructs, and we don't want those drawing as ghost rows inside
-// the card). Blank rows in the MIDDLE are preserved so paragraph
-// breaks the author put in markdown survive.
+// trimBlankBorderRows drops fully-blank edge rows from a wrapped-text
+// slice, keeping middle blanks. Logic in kit.TrimBlankBorderRows.
 func trimBlankBorderRows(rows []string) []string {
-	start := 0
-	for start < len(rows) && VisibleLen(rows[start]) == 0 {
-		start++
-	}
-	end := len(rows)
-	for end > start && VisibleLen(rows[end-1]) == 0 {
-		end--
-	}
-	if start == 0 && end == len(rows) {
-		return rows
-	}
-	return rows[start:end]
+	return kit.TrimBlankBorderRows(rows)
 }
 
-// wrapText quebra o texto em linhas que não excedem o limite.
-// Versão production-ready:
-// - Preserva quebras de linha originais
-// - Faz word-wrap por largura visível (ignora ANSI)
-// - Não destrói formatação do markdown renderizado (ANSI + linhas)
+// wrapText quebra o texto em linhas que não excedem o limite (word-wrap de
+// prosa). A matemática vive em ui/kit (kit.WrapText); este delegate preserva
+// os call sites e testes históricos do pacote.
 func wrapText(text string, limit int) []string {
-	if limit <= 0 {
-		return []string{text}
-	}
-
-	var finalLines []string
-	paragraphs := strings.Split(text, "\n")
-
-	for _, p := range paragraphs {
-		// Preserva linha vazia
-		if p == "" {
-			finalLines = append(finalLines, "")
-			continue
-		}
-
-		// Word wrap baseado em largura visível
-		words := strings.Fields(p)
-		if len(words) == 0 {
-			finalLines = append(finalLines, "")
-			continue
-		}
-
-		var line strings.Builder
-		curLen := 0
-
-		flushLine := func() {
-			finalLines = append(finalLines, line.String())
-			line.Reset()
-			curLen = 0
-		}
-
-		// emitLongWord quebra uma palavra maior que o limite em pedaços
-		// (rune-aware), empurra os pedaços completos para finalLines e
-		// deixa o último pedaço como início da linha corrente.
-		emitLongWord := func(w string) {
-			chunks := hardBreakWord(w, limit)
-			for i := 0; i < len(chunks)-1; i++ {
-				finalLines = append(finalLines, chunks[i])
-			}
-			last := chunks[len(chunks)-1]
-			line.WriteString(last)
-			curLen = VisibleLen(last)
-		}
-
-		for _, w := range words {
-			wLen := VisibleLen(w)
-			if curLen == 0 {
-				// Palavra única maior que o limite (ex.: o JSON de
-				// `last-applied-configuration` sem espaços) precisa ser
-				// quebrada aqui também — caso contrário ela é escrita
-				// inteira e estoura a largura do box.
-				if wLen > limit {
-					emitLongWord(w)
-				} else {
-					line.WriteString(w)
-					curLen = wLen
-				}
-				continue
-			}
-
-			// +1 espaço
-			if curLen+1+wLen <= limit {
-				line.WriteByte(' ')
-				line.WriteString(w)
-				curLen += 1 + wLen
-				continue
-			}
-
-			// Não cabe na linha atual: fecha e recoloca a palavra,
-			// quebrando "na marra" se ela sozinha já estoura o limite.
-			flushLine()
-
-			if wLen <= limit {
-				line.WriteString(w)
-				curLen = wLen
-				continue
-			}
-
-			emitLongWord(w)
-		}
-
-		if line.Len() > 0 {
-			finalLines = append(finalLines, line.String())
-		}
-	}
-
-	return finalLines
+	return kit.WrapText(text, limit)
 }
 
-// hardBreakWord parte uma palavra (sem espaços) em pedaços cuja largura
-// visível não excede limit. A quebra é por runa, medindo com lipgloss.Width,
-// para não cortar sequências UTF-8 / wide-runes no meio — diferente do
-// fatiamento por bytes anterior. Retorna ao menos um elemento.
-func hardBreakWord(w string, limit int) []string {
-	if limit <= 0 {
-		return []string{w}
-	}
-
-	var out []string
-	var cur strings.Builder
-	curW := 0
-	for _, rr := range w {
-		rw := lipgloss.Width(string(rr))
-		if rw < 1 {
-			rw = 1
-		}
-		if curW+rw > limit && cur.Len() > 0 {
-			out = append(out, cur.String())
-			cur.Reset()
-			curW = 0
-		}
-		cur.WriteRune(rr)
-		curW += rw
-	}
-	if cur.Len() > 0 {
-		out = append(out, cur.String())
-	}
-	if len(out) == 0 {
-		out = append(out, "")
-	}
-	return out
-}
-
-// wrapPreserve quebra texto preservando a estrutura: cada linha que cabe no
-// limite é mantida exatamente como está (indentação e espaçamento de colunas
-// intactos) e só as que estouram são quebradas, repetindo a indentação nas
-// continuações. É a mesma lógica da streaming box (wrapStreamLine), aplicada
-// linha a linha — usada no card de resultado de tool (YAML/JSON/tabelas),
-// onde colapsar whitespace como o word-wrap de prosa destruiria o layout.
+// wrapPreserve quebra texto preservando estrutura/indentação (tool output).
+// Matemática em kit.WrapPreserve.
 func wrapPreserve(text string, limit int) []string {
-	var out []string
-	for _, line := range strings.Split(text, "\n") {
-		out = append(out, wrapStreamLine(line, limit)...)
-	}
-	return out
+	return kit.WrapPreserve(text, limit)
 }
 
-// splitLeadingIndent separa a indentação inicial de uma linha do seu
-// conteúdo, de forma ANSI-aware. O glamour emite sequências de cor de
-// largura-zero ANTES (e entre) os espaços de indentação do markdown
-// renderizado — então um strings.TrimLeft(" \t") reportaria indentação
-// zero em YAML/JSON/código vindos do glamour, que é exatamente a causa do
-// bug de "perde a indentação" no modo chat. Retorna a largura visível do
-// indent em colunas, os códigos ANSI vistos na região inicial (re-anexados
-// ao conteúdo para que o primeiro token preserve a cor) e o conteúdo
-// restante.
-func splitLeadingIndent(line string) (indent int, codes string, content string) {
-	var cb strings.Builder
-	i := 0
-	for i < len(line) {
-		// Sequência CSI ANSI: preserva — pode colorir o conteúdo à frente.
-		if line[i] == 0x1b && i+1 < len(line) && line[i+1] == '[' {
-			j := i + 2
-			for j < len(line) && (line[j] < 0x40 || line[j] > 0x7e) {
-				j++
-			}
-			if j < len(line) {
-				j++ // inclui o byte final da sequência
-			}
-			cb.WriteString(line[i:j])
-			i = j
-			continue
-		}
-		if line[i] == ' ' || line[i] == '\t' {
-			indent++ // tab contado como 1 col (o glamour normaliza para espaços)
-			i++
-			continue
-		}
-		break
-	}
-	return indent, cb.String(), line[i:]
-}
-
-// wrapStructured quebra um corpo já renderizado pelo glamour para exibição
-// dentro do envelope de resposta. Diferente de wrapText (word-wrap de prosa,
-// que colapsa a indentação via strings.Fields), ele PRESERVA a indentação
-// inicial de cada linha — é o fix do bug em que YAML/JSON/código colados no
-// chat "ficam todos à esquerda e perdem a indentação". Mecânica, por linha:
-//
-//   - Detecta o indent inicial visível (ANSI-aware: o glamour emite cores
-//     ANTES dos espaços, então um TrimLeft simples não enxerga o indent).
-//   - Deslova (dedent) toda linha pela margem mínima comum — o glamour aplica
-//     uma margem de documento uniforme (tipicamente 2 cols) a prosa E código;
-//     como o próprio envelope já tem Padding(0,2), carregar a margem do
-//     glamour também empurraria tudo 2 cols para a direita.
-//   - Linhas que cabem na largura interna são emitidas verbatim (alinhamento
-//     de colunas intacto). Só as que estouram passam por word-wrap, repetindo
-//     o indent da linha em cada continuação.
+// wrapStructured quebra corpos renderizados pelo glamour preservando (e
+// dedentando) a indentação. Matemática em kit.WrapStructured.
 func wrapStructured(text string, limit int) []string {
-	if limit <= 0 {
-		return []string{text}
-	}
-	rawLines := strings.Split(text, "\n")
+	return kit.WrapStructured(text, limit)
+}
 
-	type lineSeg struct {
-		indent  int
-		payload string // códigos ANSI iniciais + conteúdo
-		blank   bool
-	}
-	segs := make([]lineSeg, 0, len(rawLines))
-	commonIndent := -1
-	for _, ln := range rawLines {
-		ind, codes, body := splitLeadingIndent(ln)
-		blank := strings.TrimSpace(stripANSIForCard(ln)) == ""
-		segs = append(segs, lineSeg{indent: ind, payload: codes + body, blank: blank})
-		if !blank && (commonIndent < 0 || ind < commonIndent) {
-			commonIndent = ind
-		}
-	}
-	if commonIndent < 0 {
-		commonIndent = 0
-	}
+// splitLeadingIndent separa indentação inicial de conteúdo, ANSI-aware.
+// Lógica em kit.SplitLeadingIndent.
+func splitLeadingIndent(line string) (indent int, codes string, content string) {
+	return kit.SplitLeadingIndent(line)
+}
 
-	out := make([]string, 0, len(rawLines))
-	for _, s := range segs {
-		if s.blank {
-			out = append(out, "")
-			continue
-		}
-		rel := s.indent - commonIndent
-		if rel < 0 {
-			rel = 0
-		}
-		pad := strings.Repeat(" ", rel)
-		full := pad + s.payload
-		if VisibleLen(full) <= limit {
-			// Cabe: verbatim — preserva qualquer alinhamento interno.
-			out = append(out, full)
-			continue
-		}
-		// Estoura: word-wrap do conteúdo, repetindo o indent nas continuações.
-		avail := limit - rel
-		if avail < 1 {
-			avail = 1
-		}
-		for _, chunk := range wrapText(s.payload, avail) {
-			out = append(out, pad+chunk)
-		}
-	}
-	return out
+// hardBreakWord parte uma palavra maior que o limite em pedaços rune-aware.
+// Lógica em kit.HardBreakWord.
+func hardBreakWord(w string, limit int) []string {
+	return kit.HardBreakWord(w, limit)
 }
 
 // RenderThinking exibe o pensamento da IA
@@ -1132,10 +764,7 @@ func cleanArgsForDisplay(args string) string {
 
 // RenderBatchHeader exibe um cabeçalho indicando o início de um lote
 func (r *UIRenderer) RenderBatchHeader(totalActions int) {
-	width, _, _ := term.GetSize(int(os.Stdout.Fd())) //#nosec G115 -- value bounded by domain
-	if width <= 0 {
-		width = 80
-	}
+	width := kit.TermWidth()
 
 	msg := i18n.T("agent.ui.batch_started", totalActions)
 	line := strings.Repeat("═", width)
@@ -1210,10 +839,7 @@ var streamBoxHeaderWidth int
 func (r *UIRenderer) RenderStreamBoxStart(icon, title, color string) {
 	header := fmt.Sprintf("%s %s", icon, title)
 
-	termWidth, _, err := term.GetSize(int(os.Stdout.Fd())) //#nosec G115 -- value bounded by domain
-	if err != nil || termWidth <= 0 {
-		termWidth = 80
-	}
+	termWidth := kit.TermWidth()
 
 	// "╭── " (4) + header + " " + minimum 6 dashes for visual weight.
 	const minTail = 6
