+++ cli/ui_boxes_test.go
@@ -3,7 +3,18 @@ package cli
 import (
 	"strings"
 	"testing"
+
+	"github.com/diillson/chatcli/ui/theme"
 )
+
+// pinANSIProfile fixes the color profile so truncation assertions about the
+// trailing reset are deterministic (kit.Truncate emits escapes only on
+// colored profiles), restoring detection afterwards.
+func pinANSIProfile(t *testing.T) {
+	t.Helper()
+	theme.SetProfile(theme.ProfileANSI)
+	t.Cleanup(func() { theme.SetProfile(theme.DetectProfile()) })
+}
 
 func TestAnsiTruncate_ShortStringUnchanged(t *testing.T) {
 	in := "\033[33mshort\033[0m line"
@@ -13,6 +24,7 @@ func TestAnsiTruncate_ShortStringUnchanged(t *testing.T) {
 }
 
 func TestAnsiTruncate_ClampsVisibleWidth(t *testing.T) {
+	pinANSIProfile(t)
 	in := "\033[35m[servicenow/incidents]\033[0m " + strings.Repeat("x", 200)
 	got := ansiTruncate(in, 40)
 
@@ -28,6 +40,7 @@ func TestAnsiTruncate_ClampsVisibleWidth(t *testing.T) {
 }
 
 func TestAnsiTruncate_WideRunesCountPerCell(t *testing.T) {
+	pinANSIProfile(t)
 	// Emoji and CJK occupy two terminal columns; byte- or rune-based
 	// truncation would overflow the box and wrap the line.
 	in := strings.Repeat("🚀", 30)
@@ -43,9 +56,10 @@ func TestAnsiTruncate_WideRunesCountPerCell(t *testing.T) {
 func TestUIBoxLine_NeverExceedsTerminalWidth(t *testing.T) {
 	long := strings.Repeat("word ", 100)
 	got := uiBoxLine(ColorPurple, long)
-	// uiTermWidth falls back to 80 outside a TTY (the test environment).
-	if w := visibleLen(got); w > 80 {
-		t.Errorf("box line visible width = %d, want <= 80", w)
+	// uiTermWidth delegates to kit.TermWidth, which falls back to 100
+	// outside a TTY (the test environment) — the unified fallback.
+	if w := visibleLen(got); w > uiTermWidth() {
+		t.Errorf("box line visible width = %d, want <= %d", w, uiTermWidth())
 	}
 	if !strings.Contains(got, "│") {
 		t.Error("box line must keep the sidebar prefix")
